app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich>=0.9.2",
    "fastapi>=0.120.3",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "pywebview>=6.1",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.38.0",
]

[tool.aerich]
tortoise_orm = "DBsettings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
